+++ CHANGELOG.md
@@ -12,6 +12,8 @@ New Features:
 
 - And one for the hackers: File logging is now additionally available in the extension process of VS Code and not only in the debug adapter process. To activate this feature you have to add `"vscode-janus-debug.log"` to your settings ([#30](https://github.com/otris/vscode-janus-debug/issues/30)).
 
+- The configuration name for the extension settings in `settings.json` is now `"vscode-janus-debug"` instead of `"vscode-documents-scripting"`. The user has to change the name for any list in `settings.json`.
+
 ### 0.0.8 (2017-07-07)
 
 New Features: